Added check permissions, getting location, animation in menu, showing current weather, dialog, formatting, styles

+++ models/src/main/java/ru/gc986/models/weather/Main.kt
@@ -2,8 +2,8 @@ package ru.gc986.models.weather
 
 data class Main(
     val feels_like: Double,
-    val humidity: Int,
-    val pressure: Int,
+    val humidity: Double,
+    val pressure: Double,
     val temp: Double,
     val temp_max: Double,
     val temp_min: Double
